Move rlp utils to dedicated file
